#8.13-#8.15

+++ src/controllers/videosControllers.js
@@ -13,7 +13,6 @@ export const home = async(req, res) => {
 export const watch = async (req, res) => {
     const { id } = req.params;
     const video = await Video.findById(id).populate("owner");
-    console.log(video);
     if(!video) {
         return res.render("404", { pageTitle: "Video not Found."});
     }
@@ -24,6 +23,9 @@ export const getEdit = async (req, res) => {
     const video = await Video.findById(id);
     if(!video) {
         return res.status(404).render("404", { pageTitle: "Video not Found."});
+    }
+    if (String(video.owner) !== req.session.user._id) {
+        return res.status(403).redirect("/");
     }
     return res.render("edit", { pageTitle: `Editing ${video.title}`, video});
 };
@@ -42,6 +44,9 @@ export const postEdit = async (req, res) => {
     } catch (error) {
         return res.status(404).render("404", { pageTitle: "Edited video saving fail" });
     }
+    if (String(videoExist.owner) !== req.session.user._id) {
+        return res.status(403).redirect("/");
+    }
     return res.redirect(`/videos/${id}`);
 };
 export const getUpload = (req, res) => {
@@ -52,22 +57,32 @@ export const postUpload = async (req, res) => {
     const file=req.file;
     const { title, description, hashtags } = req.body;
     try {
-        await Video.create({
+        const newVideo = await Video.create({
             title,
             description,
             fileUrl: file ? file.path : "",
             hashtags: Video.formatHashtags(hashtags),
             owner: _id,
         });
-        return res.redirect("/");
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save();
+       return res.redirect("/");
     } catch(error) {
         return res.status(400).render("upload", { pageTitle: "Upload Video", errorMessage: error._message });
     }
 };
 
-
 export const getDelete = async (req, res) => {
     const { id } = req.params;
+    const video = await Video.findById(id);
+    if (!video) {
+        return res.render("404", { pageTitle: "Video not found." });
+    }
+    if (String(video.owner) !== req.session.user._id) {
+        return res.status(403).redirect("/");
+        
+    }
     await Video.findByIdAndDelete(id);
     return res.redirect("/");
 };
